Use @handlewithcare/react-codemirror in CodeBlock
@handlewithcare/react-codemirror has some behavior
over @uiw/react-codemirror that makes it better suited
for a ProseMirror integration. It allows you to lift the
EditorState out of the CodeMirror component, which means
that we can derive parts of the state, like the selection
from the ProseMirror state.

We can also compute smaller diffs when updating the state,
rather than completely replacing the doc (which triggers
recalculation of the syntax highlighting, decorations, etc)
any time the node content is updated from outside the
CodeMirror editor.